+++ src/repeat/RepeatMaskerLine.java
@@ -8,13 +8,13 @@ import edu.princeton.cs.introcs.Out;
  * Date: 11.03.14
  */
 public class RepeatMaskerLine {
-    public long swScore;
+    public int swScore;
     public double percDiv;
     public double percDel;
     public double percIns;
     public String querySeq;
-    public long posQBegin;
-    public long posQEnd;
+    public int posQBegin;
+    public int posQEnd;
     public String posQLeft;
     public String complement;
     public String repeatName;
@@ -22,7 +22,7 @@ public class RepeatMaskerLine {
     public String posRBegin;
     public String posREnd;
     public String posRLeft;
-    public long id;
+    public int id;
     public boolean asteriks;
 
     public RepeatMaskerLine() {
@@ -41,13 +41,13 @@ public class RepeatMaskerLine {
             RepeatMaskerLine rml = new RepeatMaskerLine();
             line = line.trim();
             String[] fields = line.split("\\s+");
-            rml.swScore = Long.parseLong(fields[0]);
+            rml.swScore = Integer.parseInt(fields[0]);
             rml.percDiv = Double.parseDouble(fields[1]);
             rml.percDel = Double.parseDouble(fields[2]);
             rml.percIns = Double.parseDouble(fields[3]);
             rml.querySeq = fields[4];
-            rml.posQBegin = Long.parseLong(fields[5]);
-            rml.posQEnd = Long.parseLong(fields[6]);
+            rml.posQBegin = Integer.parseInt(fields[5]);
+            rml.posQEnd = Integer.parseInt(fields[6]);
             rml.posQLeft = fields[7];
             rml.complement = fields[8];
             rml.repeatName = fields[9];
@@ -55,7 +55,7 @@ public class RepeatMaskerLine {
             rml.posRBegin = fields[11];
             rml.posREnd = fields[12];
             rml.posRLeft = fields[13];
-            rml.id = Long.parseLong(fields[14]);
+            rml.id = Integer.parseInt(fields[14]);
             rml.asteriks = (fields.length == 16 && fields[15].equals("*"));
             return rml;
         } catch (Exception e) {
